Remove deprecated classes
-Replace ComputeProvider with ComputeProvider2.
-Delete ComputeProvider2.
-Delete ComputeMetricGroup and old widgets.
-Remove csv parser from view project.
