+++ src/view/src/rocprofvis_compute_root.cpp
@@ -2,7 +2,6 @@
 
 #include "rocprofvis_compute_root.h"
 #include "rocprofvis_compute_data_provider.h"
-#include "rocprofvis_compute_data_provider2.h"
 #include "rocprofvis_compute_block.h"
 #include "rocprofvis_compute_roofline.h"
 #include "rocprofvis_compute_summary.h"
@@ -17,21 +16,20 @@ namespace View
 
 ComputeRoot::ComputeRoot()
 : m_compute_data_provider(nullptr)
-, m_compute_data_provider2(nullptr)
 , m_tab_container(nullptr)
 , m_data_dirty_event_token(-1)
 , m_data_dirty(false)
+, m_trace_opened(false)
 {
     m_id = GenUniqueName("");
     
     m_compute_data_provider = std::make_shared<ComputeDataProvider>();
-    m_compute_data_provider2 = std::make_shared<ComputeDataProvider2>();
 
     m_tab_container = std::make_shared<TabContainer>();
-    m_tab_container->AddTab(TabItem{"Summary View", COMPUTE_SUMMARY_VIEW_URL, std::make_shared<ComputeSummaryView>(m_id, m_compute_data_provider2), false});
+    m_tab_container->AddTab(TabItem{"Summary View", COMPUTE_SUMMARY_VIEW_URL, std::make_shared<ComputeSummaryView>(m_id, m_compute_data_provider), false});
     m_tab_container->AddTab(TabItem{"Roofline View", COMPUTE_ROOFLINE_VIEW_URL, std::make_shared<ComputeRooflineView>(m_id, m_compute_data_provider), false});
-    m_tab_container->AddTab(TabItem{"Block View", COMPUTE_BLOCK_VIEW_URL, std::make_shared<ComputeBlockView>(m_id, m_compute_data_provider2), false});
-    m_tab_container->AddTab(TabItem{"Table View", COMPUTE_TABLE_VIEW_URL, std::make_shared<ComputeTableView>(m_id, m_compute_data_provider2), false});
+    m_tab_container->AddTab(TabItem{"Block View", COMPUTE_BLOCK_VIEW_URL, std::make_shared<ComputeBlockView>(m_id, m_compute_data_provider), false});
+    m_tab_container->AddTab(TabItem{"Table View", COMPUTE_TABLE_VIEW_URL, std::make_shared<ComputeTableView>(m_id, m_compute_data_provider), false});
 
     NavigationManager::GetInstance()->RegisterContainer(m_tab_container, m_id, m_id);
 
@@ -42,25 +40,14 @@ ComputeRoot::ComputeRoot()
     m_data_dirty_event_token = EventManager::GetInstance()->Subscribe(static_cast<int>(RocEvents::kComputeDataDirty), data_dirty_event_handler);
 }
 
-ComputeRoot::~ComputeRoot() {
+ComputeRoot::~ComputeRoot() 
+{
     NavigationManager::GetInstance()->UnregisterContainer(m_tab_container, m_id, m_id);
 }
 
 void ComputeRoot::Update()
 {
-    if (!m_compute_data_provider->ProfileLoaded())
-    {
-        m_compute_data_provider->LoadProfile();
-
-        if (m_compute_data_provider->ProfileLoaded())
-        {
-            if (m_tab_container)
-            {
-                m_tab_container->Update();
-            }
-        }
-    }
-    if (m_data_dirty)
+    if (m_data_dirty && m_trace_opened)
     {
         m_data_dirty = false;
         if (m_tab_container)
@@ -70,24 +57,14 @@ void ComputeRoot::Update()
     }
 }
 
-void ComputeRoot::SetProfilePath(const std::string& path)
-{
-    m_compute_data_provider->SetProfilePath(path);
-}
-
 void ComputeRoot::OpenTrace(const std::string& path)
 {
-    m_compute_data_provider2->LoadTrace(path);
-}
-
-bool ComputeRoot::ProfileLoaded()
-{
-    return m_compute_data_provider->ProfileLoaded();
+    m_trace_opened = (kRocProfVisResultSuccess == m_compute_data_provider->LoadTrace(path));
 }
 
 void ComputeRoot::Render()
 {
-    if (m_compute_data_provider->ProfileLoaded())
+    if (m_trace_opened)
     {
         if (m_tab_container)
         {
