+++ src/view/src/rocprofvis_compute_root.h
@@ -10,7 +10,6 @@ namespace View
 {
 
 class ComputeDataProvider;
-class ComputeDataProvider2;
 
 class ComputeRoot : public RocWidget
 {
@@ -18,18 +17,16 @@ public:
     void Render() override;
     void Update() override;
     void OpenTrace(const std::string& path);
-    void SetProfilePath(const std::string& path);
-    bool ProfileLoaded();
     ComputeRoot();
     ~ComputeRoot();
 
 private:
     std::shared_ptr<TabContainer> m_tab_container;
     std::shared_ptr<ComputeDataProvider> m_compute_data_provider;
-    std::shared_ptr<ComputeDataProvider2> m_compute_data_provider2;
     std::string m_id;
     EventManager::SubscriptionToken m_data_dirty_event_token;
     bool m_data_dirty;
+    bool m_trace_opened;
 };
 
 }  // namespace View
